Add reset.css for reference
Added the reset.css file to the project to reference the source of the CSS reset from Piccalilli's Modern CSS Reset. While reset.min.css will be used for production to minimize file size, reset.css serves as a reference for developers to view and understand the source CSS reset rules.